airflow-dbt/requirements.txt
apache-airflow-providers-amazon>=9.12.0
apache-airflow-providers-standard>=1.6.0
apache-airflow-task-sdk>=1.0.4
astronomer-cosmos==1.10.2
dbt-core==1.10.9
dbt-postgres==1.9.0
dbt-redshift>=1.9.5
opentelemetry-api>=1.28.0
opentelemetry-exporter-otlp>=1.28.0
